refactor(service): split manager by responsibility
