+++ internal/service/manager.go
@@ -5,12 +5,9 @@ import (
 	"errors"
 	"fmt"
 	"reflect"
-	"regexp"
 	"sort"
-	"strings"
 	"sync"
 	"sync/atomic"
-	"syscall"
 	"time"
 
 	"github.com/kranz-org/kranz/internal/config"
@@ -269,1101 +266,7 @@ func (m *Manager) configSnapshot() *config.Config {
 }
 
 // StartService starts one service after validating ports and dependencies.
-func (m *Manager) StartService(name string) error {
-	return m.startService(context.Background(), name, false)
-}
 
-func (m *Manager) startService(ctx context.Context, name string, recovery bool) error {
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	if m.shuttingDown.Load() {
-		return errors.New("application is shutting down; new processes are disabled")
-	}
-	svc, ok := m.GetService(name)
-	if !ok {
-		return fmt.Errorf("service %q not found", name)
-	}
-	svc.lifecycleMu.Lock()
-	defer svc.lifecycleMu.Unlock()
-	if svc.Config.IsDetached() {
-		if alreadyRunning(svc.Status()) {
-			return m.startDetachedService(ctx, svc)
-		}
-		if err := m.runPrerequisites(ctx, svc); err != nil {
-			return err
-		}
-		return m.startDetachedService(ctx, svc)
-	}
-
-	if status := svc.Status(); status != config.StatusStopped {
-		return fmt.Errorf("service %q is already running", name)
-	}
-	// Prerequisites run after dependencies are ready and before this service
-	// starts, so a failed prerequisite leaves the service stopped rather than
-	// starting it against an unprepared environment.
-	if err := m.runPrerequisites(ctx, svc); err != nil {
-		return err
-	}
-	svc.SetDesiredRunning(true)
-	if !recovery {
-		svc.ResetRestartCount()
-	}
-	if m.shuttingDown.Load() {
-		svc.SetDesiredRunning(false)
-		return errors.New("application is shutting down; new processes are disabled")
-	}
-
-	// Pre-flight port check
-	m.mu.RLock()
-	pc := m.portChecker
-	m.mu.RUnlock()
-	if pc != nil && len(svc.Config.Ports) > 0 {
-		portsInfo, err := pc.CheckPorts(svc.Config.Ports)
-		if err == nil {
-			for _, port := range svc.Config.Ports {
-				if info, ok := portsInfo[port]; ok && info != nil {
-					owner := m.ManagedServiceForPID(info.PID)
-					svc.SetDesiredRunning(false)
-					return &PortConflictError{
-						Service:      name,
-						Port:         port,
-						PID:          info.PID,
-						Process:      info.Process,
-						Command:      info.Command,
-						OwnerService: owner,
-						External:     owner == "",
-					}
-				}
-			}
-		}
-	}
-
-	svc.SetStatus(config.StatusStarting)
-	svc.AppendLog("[Kranz] Starting")
-
-	pm := NewProcessManager(1000)
-	start := svc.Config.StartAction()
-	if start == nil {
-		svc.SetDesiredRunning(false)
-		svc.SetStatus(config.StatusStopped)
-		return fmt.Errorf("service %q has no start capability", name)
-	}
-
-	pid, err := pm.Start(ctx, start.Command, start.Dir, start.Env, start.Shell)
-	if err != nil {
-		svc.SetDesiredRunning(false)
-		svc.SetStatus(config.StatusStopped)
-		svc.AppendLog("[Kranz] Start failed: " + err.Error())
-		return fmt.Errorf("start service %q: %w", name, err)
-	}
-
-	svc.SetPID(pid)
-	svc.SetStatus(config.StatusRunning)
-	svc.AppendLog(fmt.Sprintf("[Kranz] Started · PID %d", pid))
-	svc.ResetNewLogCount()
-
-	monitorStop := make(chan struct{})
-	svc.setRuntime(pm, monitorStop)
-	if svc.Config.PortDiscoveryEnabled() {
-		m.ensureListenerDiscovery()
-	}
-
-	m.mu.RLock()
-	hc := m.healthChecker
-	m.mu.RUnlock()
-	if hc != nil {
-		hc.StartMonitoring(name, svc.Config.HealthCheck)
-	}
-
-	// Output monitoring owns process completion and recovery transitions.
-	go m.monitorProcess(name, svc, pm, monitorStop)
-
-	return nil
-}
-
-// alreadyRunning reports states in which a detached start is a no-op, so that
-// prerequisites are not run again for a resource that is already up.
-func alreadyRunning(status config.ServiceStatus) bool {
-	return status == config.StatusRunning || status == config.StatusUnhealthy
-}
-
-func (m *Manager) startDetachedService(ctx context.Context, svc *Service) error {
-	start := svc.Config.Lifecycle.Start
-	if start == nil {
-		return fmt.Errorf("service %q has no start capability", svc.Name)
-	}
-	status := svc.Status()
-	if status == config.StatusRunning || status == config.StatusUnhealthy {
-		return nil
-	}
-	if status != config.StatusStopped && status != config.StatusUnknown {
-		return fmt.Errorf("service %q is already running", svc.Name)
-	}
-	if status == config.StatusUnknown && svc.Config.Lifecycle.Status != nil {
-		failures := 0
-		m.reconcileDetachedStatus(ctx, svc, &failures)
-		if ctx.Err() != nil {
-			return ctx.Err()
-		}
-		if observed := svc.Status(); observed == config.StatusRunning || observed == config.StatusUnhealthy {
-			return nil
-		}
-	}
-	svc.SetDesiredRunning(true)
-	svc.SetStatus(config.StatusStarting)
-	svc.AppendLog("[Kranz] Starting detached service")
-	id := lifecycleActionID(svc.Name, "start")
-	result, err := m.actions.RunDefinition(ctx, id, *start)
-	m.appendLifecycleResult(svc, "start", result)
-	if err != nil {
-		svc.SetDesiredRunning(false)
-		// A detached start can mutate external state before failing, timing out,
-		// or being canceled. Never claim it is stopped without observing that.
-		svc.SetStatus(config.StatusUnknown)
-		return fmt.Errorf("start detached service %q: %w", svc.Name, err)
-	}
-	svc.SetPID(0)
-	svc.SetStatus(config.StatusRunning)
-	m.mu.RLock()
-	hc := m.healthChecker
-	m.mu.RUnlock()
-	if hc != nil {
-		hc.StartMonitoring(svc.Name, svc.Config.HealthCheck)
-	}
-	m.startDetachedLogs(svc)
-	m.wakeStatusMonitor(svc.Name)
-	return nil
-}
-
-// StopService gracefully stops one service and releases its process group.
-func (m *Manager) StopService(name string) error {
-	svc, ok := m.GetService(name)
-	if !ok {
-		return fmt.Errorf("service %q not found", name)
-	}
-	svc.lifecycleMu.Lock()
-	defer svc.lifecycleMu.Unlock()
-	if svc.Config.IsDetached() {
-		return m.stopDetachedService(svc)
-	}
-	svc.SetDesiredRunning(false)
-
-	pm, monitorStop := svc.runtime()
-	if svc.Status() == config.StatusStopped && pm == nil {
-		return nil
-	}
-
-	svc.SetStatus(config.StatusStopping)
-	svc.AppendLog("[Kranz] Stopping")
-
-	if monitorStop != nil {
-		close(monitorStop)
-	}
-
-	m.mu.RLock()
-	hc := m.healthChecker
-	m.mu.RUnlock()
-	if hc != nil {
-		hc.StopMonitoring(name)
-	}
-
-	var stopErr error
-	if pm != nil {
-		shutdown := svc.Config.Shutdown
-		stopErr = pm.StopWithOptions(StopOptions{
-			Command: shutdown.Command, Timeout: shutdown.Timeout, Signal: syscall.Signal(shutdown.Signal),
-			ParentOnly: shutdown.ParentOnly, Dir: svc.Config.Dir, Env: svc.Config.Env, Shell: svc.Config.Shell,
-		})
-		m.drainProcessLogs(svc, pm)
-		svc.clearRuntime(pm)
-	}
-	if stopErr != nil {
-		svc.AppendLog(fmt.Sprintf("[Kranz] Failed to stop %s: %v", name, stopErr))
-	}
-
-	svc.SetPID(0)
-	svc.SetStatus(config.StatusStopped)
-	svc.AppendLog("[Kranz] Stopped")
-	return stopErr
-}
-
-func (m *Manager) stopDetachedService(svc *Service) error {
-	if svc.Status() == config.StatusStopped {
-		return nil
-	}
-	stop := svc.Config.Lifecycle.Stop
-	if stop == nil {
-		return fmt.Errorf("service %q has no stop capability", svc.Name)
-	}
-	svc.SetDesiredRunning(false)
-	svc.SetStatus(config.StatusStopping)
-	svc.AppendLog("[Kranz] Stopping detached service")
-	m.stopDetachedLogs(svc.Name)
-	m.mu.RLock()
-	hc := m.healthChecker
-	m.mu.RUnlock()
-	if hc != nil {
-		hc.StopMonitoring(svc.Name)
-	}
-	id := lifecycleActionID(svc.Name, "stop")
-	result, err := m.actions.RunDefinition(context.Background(), id, *stop)
-	m.appendLifecycleResult(svc, "stop", result)
-	if err != nil {
-		svc.SetStatus(config.StatusUnknown)
-		m.wakeStatusMonitor(svc.Name)
-		return fmt.Errorf("stop detached service %q: %w", svc.Name, err)
-	}
-	svc.SetPID(0)
-	svc.SetStatus(config.StatusStopped)
-	m.wakeStatusMonitor(svc.Name)
-	return nil
-}
-
-func lifecycleActionID(serviceName, operation string) config.ActionID {
-	return config.ActionID{OwnerKind: config.ActionOwnerLifecycle, Owner: serviceName, Name: operation}
-}
-
-func (m *Manager) appendLifecycleResult(svc *Service, operation string, result ActionResult) {
-	// Compose and similar tools render progress to stderr even on success. A
-	// successful lifecycle operation only needs its concise boundary; keep
-	// command output for failures where it is diagnostic.
-	if result.Status != ActionSucceeded {
-		lines := append([]string(nil), result.Stdout...)
-		for _, line := range result.Stderr {
-			lines = append(lines, "[stderr] "+line)
-		}
-		const maxLifecycleFailureLines = 40
-		if omitted := len(lines) - maxLifecycleFailureLines; omitted > 0 {
-			svc.AppendLog(fmt.Sprintf("[Kranz] %d earlier lifecycle output lines omitted", omitted))
-			lines = lines[omitted:]
-		}
-		for _, line := range lines {
-			svc.AppendLog(line)
-		}
-	}
-	var resultErr error
-	if result.Error != "" {
-		resultErr = errors.New(result.Error)
-	}
-	svc.RecordExit(result.ExitCode, resultErr)
-	svc.AppendLog(fmt.Sprintf("[Kranz] Lifecycle %s %s · exit %d", operation, result.Status.String(), result.ExitCode))
-}
-
-// StartAll starts every enabled service in dependency order.
-func (m *Manager) StartAll() error {
-	return m.StartServices(m.enabledServiceNames())
-}
-
-// enabledServiceNames lists the services a "start everything" operation may
-// touch. A disabled service is excluded: it remains startable by name, which is
-// the difference between hidden and manual.
-func (m *Manager) enabledServiceNames() []string {
-	cfg := m.configSnapshot()
-	names := make([]string, 0, len(cfg.Services))
-	for _, name := range cfg.ServiceNames() {
-		if !cfg.Services[name].Disabled {
-			names = append(names, name)
-		}
-	}
-	return names
-}
-
-// StartAllContext starts all services and lets callers cancel readiness waits.
-func (m *Manager) StartAllContext(ctx context.Context) error {
-	return m.StartServicesContext(ctx, m.enabledServiceNames())
-}
-
-// StartServices starts the requested services and any dependencies they require.
-// Services outside that dependency closure are left untouched.
-func (m *Manager) StartServices(names []string) error {
-	return m.StartServicesContext(context.Background(), names)
-}
-
-// ForceStartServices starts exactly the requested stopped services without
-// expanding or waiting for dependencies. Normal port and process ownership
-// checks still apply.
-func (m *Manager) ForceStartServices(names []string) error {
-	unique := make([]string, 0, len(names))
-	seen := make(map[string]bool, len(names))
-	for _, name := range names {
-		if seen[name] {
-			continue
-		}
-		if _, ok := m.GetService(name); !ok {
-			return fmt.Errorf("service %q not found", name)
-		}
-		seen[name] = true
-		unique = append(unique, name)
-	}
-
-	var startErrors []error
-	for _, name := range unique {
-		svc, _ := m.GetService(name)
-		if !svc.CanStart() {
-			continue
-		}
-		if err := m.StartService(name); err != nil {
-			wrapped := fmt.Errorf("%s: %w", name, err)
-			startErrors = append(startErrors, wrapped)
-			svc.AppendLog(fmt.Sprintf("[Kranz] Force start failed: %v", err))
-		}
-	}
-	return errors.Join(startErrors...)
-}
-
-// StartServicesContext starts a dependency closure and stops launching new
-// processes as soon as the context is canceled.
-func (m *Manager) StartServicesContext(ctx context.Context, names []string) error {
-	order, err := m.topologicalSort()
-	if err != nil {
-		return err
-	}
-	selected, err := m.expandWithDependencies(names)
-	if err != nil {
-		return err
-	}
-	queued := m.queuePendingStarts(selected)
-	defer m.clearPendingStarts(queued)
-
-	groups := m.groupByDependencyLevel(order)
-	var startErrors []error
-
-	for _, group := range groups {
-		if err := ctx.Err(); err != nil {
-			return errors.Join(errors.Join(startErrors...), err)
-		}
-		groupErrors, err := m.startDependencyGroup(ctx, group, selected)
-		startErrors = append(startErrors, groupErrors...)
-		if err != nil {
-			return errors.Join(errors.Join(startErrors...), err)
-		}
-		if err := m.waitForDependencyGroup(ctx, group, selected); err != nil {
-			return errors.Join(errors.Join(startErrors...), err)
-		}
-	}
-
-	return errors.Join(startErrors...)
-}
-
-type pendingStartIntent struct {
-	name      string
-	startedAt time.Time
-}
-
-// queuePendingStarts exposes the complete dependency closure before the first
-// dependency gate blocks. DesiredRunning is already the lifecycle source of
-// truth, so the UI can render this intent without inventing another state.
-func (m *Manager) queuePendingStarts(selected map[string]bool) []pendingStartIntent {
-	queued := make([]pendingStartIntent, 0, len(selected))
-	for name := range selected {
-		svc, ok := m.GetService(name)
-		if !ok || !svc.CanStart() {
-			continue
-		}
-		startedAt := svc.GetState().StartedAt
-		svc.SetDesiredRunning(true)
-		queued = append(queued, pendingStartIntent{name: name, startedAt: startedAt})
-	}
-	return queued
-}
-
-func (m *Manager) clearPendingStarts(queued []pendingStartIntent) {
-	for _, intent := range queued {
-		svc, ok := m.GetService(intent.name)
-		if !ok {
-			continue
-		}
-		state := svc.GetState()
-		if (state.Status == config.StatusStopped || state.Status == config.StatusUnknown) && state.StartedAt.Equal(intent.startedAt) {
-			svc.SetDesiredRunning(false)
-		}
-	}
-}
-
-func (m *Manager) startDependencyGroup(ctx context.Context, group []string, selected map[string]bool) ([]error, error) {
-	var startErrors []error
-	for _, name := range group {
-		if !selected[name] {
-			continue
-		}
-		if err := ctx.Err(); err != nil {
-			return startErrors, err
-		}
-		svc, _ := m.GetService(name)
-		if svc != nil && !svc.CanStart() {
-			continue
-		}
-		if err := m.startService(ctx, name, false); err != nil {
-			startErrors = append(startErrors, fmt.Errorf("%s: %w", name, err))
-			if svc, _ := m.GetService(name); svc != nil {
-				svc.AppendLog(fmt.Sprintf("[Kranz] Start failed: %v", err))
-			}
-		}
-	}
-	return startErrors, nil
-}
-
-func (m *Manager) waitForDependencyGroup(ctx context.Context, group []string, selected map[string]bool) error {
-	for _, dependencyName := range group {
-		if !selected[dependencyName] {
-			continue
-		}
-		conditions := m.conditionsForDependency(dependencyName, selected)
-		for _, condition := range conditions {
-			if err := m.waitForDependencyCondition(ctx, dependencyName, condition); err != nil {
-				m.handleSkippedDependents(dependencyName, selected)
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-func (m *Manager) conditionsForDependency(dependencyName string, selected map[string]bool) []config.DependencyCondition {
-	seen := make(map[config.DependencyCondition]bool)
-	var result []config.DependencyCondition
-	for _, dependent := range m.Services() {
-		if !selected[dependent.Name] || !containsName(dependent.Config.DependsOn, dependencyName) {
-			continue
-		}
-		condition := config.DependencyHealthy
-		if configured, ok := dependent.Config.DependencyConditions[dependencyName]; ok && configured.Condition != "" {
-			condition = configured.Condition
-		}
-		if !seen[condition] {
-			seen[condition] = true
-			result = append(result, condition)
-		}
-	}
-	return result
-}
-
-func (m *Manager) waitForDependencyCondition(ctx context.Context, name string, condition config.DependencyCondition) error {
-	svc, ok := m.GetService(name)
-	if !ok {
-		return fmt.Errorf("dependency %s disappeared", name)
-	}
-	if condition == config.DependencyHealthy {
-		if m.waitForReadiness(ctx, name, 30*time.Second) {
-			return nil
-		}
-		if err := ctx.Err(); err != nil {
-			return err
-		}
-		return fmt.Errorf("dependency %s did not become healthy within 30s", name)
-	}
-	var readyPattern *regexp.Regexp
-	if condition == config.DependencyLogReady {
-		var err error
-		readyPattern, err = regexp.Compile(svc.Config.ReadyLogLine)
-		if err != nil {
-			return fmt.Errorf("dependency %s ready_log_line: %w", name, err)
-		}
-	}
-	ticker := time.NewTicker(50 * time.Millisecond)
-	defer ticker.Stop()
-	startedDeadline := time.NewTimer(30 * time.Second)
-	defer startedDeadline.Stop()
-	for {
-		state := svc.GetState()
-		switch condition {
-		case config.DependencyStarted:
-			if !state.StartedAt.IsZero() && state.Status != config.StatusStopped && state.Status != config.StatusUnknown {
-				return nil
-			}
-		case config.DependencyCompleted:
-			if state.Completed {
-				return nil
-			}
-		case config.DependencyCompletedSuccessfully:
-			if state.Completed {
-				if m.successfulExit(svc.Config, state.ExitCode) {
-					return nil
-				}
-				return fmt.Errorf("dependency %s completed with exit code %d", name, state.ExitCode)
-			}
-		case config.DependencyLogReady:
-			for _, line := range svc.Logs.Lines() {
-				if readyPattern.MatchString(line) {
-					return nil
-				}
-			}
-		default:
-			return fmt.Errorf("dependency %s has unsupported condition %q", name, condition)
-		}
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-ticker.C:
-		case <-startedDeadline.C:
-			if condition == config.DependencyStarted {
-				return fmt.Errorf("dependency %s did not start within 30s", name)
-			}
-			startedDeadline.Reset(24 * time.Hour)
-		}
-	}
-}
-
-func (m *Manager) handleSkippedDependents(dependencyName string, selected map[string]bool) {
-	for _, dependent := range m.Services() {
-		if selected[dependent.Name] && containsName(dependent.Config.DependsOn, dependencyName) && dependent.Config.Availability.ExitOnSkipped {
-			m.requestProjectExit(1)
-			return
-		}
-	}
-}
-
-func containsName(values []string, target string) bool {
-	for _, value := range values {
-		if value == target {
-			return true
-		}
-	}
-	return false
-}
-
-// groupByDependencyLevel groups independent services for parallel readiness gating.
-func (m *Manager) groupByDependencyLevel(order []string) [][]string {
-	graph := m.configSnapshot().GetDependsOn()
-	levels := make(map[string]int)
-
-	// A service level is one more than its deepest dependency.
-	for _, name := range order {
-		level := 0
-		for _, dep := range graph[name] {
-			if levels[dep] >= level {
-				level = levels[dep] + 1
-			}
-		}
-		levels[name] = level
-	}
-
-	// Pre-size the level buckets from the deepest service.
-	maxLevel := 0
-	for _, lvl := range levels {
-		if lvl > maxLevel {
-			maxLevel = lvl
-		}
-	}
-
-	// Services within one level have no ordering dependency on each other.
-	groups := make([][]string, maxLevel+1)
-	for i := range groups {
-		groups[i] = make([]string, 0)
-	}
-	for _, name := range order {
-		lvl := levels[name]
-		groups[lvl] = append(groups[lvl], name)
-	}
-
-	return groups
-}
-
-// StartByTags starts services matching at least one tag and their dependencies.
-func (m *Manager) StartByTags(tags []string) error {
-	names := m.configSnapshot().GetServicesByTags(tags)
-	if len(names) == 0 {
-		return fmt.Errorf("no services match tags %v", tags)
-	}
-
-	var startErrors []error
-	for _, name := range names {
-		if err := m.StartService(name); err != nil {
-			startErrors = append(startErrors, fmt.Errorf("%s: %w", name, err))
-			svc, _ := m.GetService(name)
-			if svc != nil {
-				svc.AppendLog(fmt.Sprintf("[Kranz] Start failed: %v", err))
-			}
-		}
-	}
-	return errors.Join(startErrors...)
-}
-
-// StopAll stops every service in reverse dependency order.
-func (m *Manager) StopAll() error {
-	return m.stopServices(m.configSnapshot().ServiceNames(), false)
-}
-
-// StopServices stops the requested services and every transitive dependent, in
-// reverse dependency order.
-func (m *Manager) StopServices(names []string) error {
-	return m.stopServices(names, true)
-}
-
-// ForceStopServices stops exactly the requested services without expanding
-// dependents. It is the shutdown counterpart to ForceStartServices.
-func (m *Manager) ForceStopServices(names []string) error {
-	return m.stopServices(names, false)
-}
-
-func (m *Manager) stopServices(names []string, includeDependents bool) error {
-	order, err := m.topologicalSort()
-	if err != nil {
-		order = m.configSnapshot().ServiceNames()
-	}
-	selected := make(map[string]bool, len(names))
-	for _, name := range names {
-		if _, ok := m.GetService(name); !ok {
-			return fmt.Errorf("service %q not found", name)
-		}
-		selected[name] = true
-	}
-	if includeDependents {
-		for _, name := range names {
-			for _, dependent := range m.findDependents(name) {
-				selected[dependent] = true
-			}
-		}
-	}
-
-	var stopErrors []error
-	for i := len(order) - 1; i >= 0; i-- {
-		name := order[i]
-		if !selected[name] {
-			continue
-		}
-		if svc, ok := m.GetService(name); ok && svc.Config.IsDetached() && !svc.CanStop() {
-			continue
-		}
-		if err := m.StopService(name); err != nil {
-			stopErrors = append(stopErrors, fmt.Errorf("%s: %w", name, err))
-			svc, _ := m.GetService(name)
-			if svc != nil {
-				svc.AppendLog(fmt.Sprintf("[Kranz] Stop failed: %v", err))
-			}
-		}
-	}
-
-	return errors.Join(stopErrors...)
-}
-
-func (m *Manager) expandWithDependencies(names []string) (map[string]bool, error) {
-	selected := make(map[string]bool, len(names))
-	var visit func(string) error
-	visit = func(name string) error {
-		if selected[name] {
-			return nil
-		}
-		svc, ok := m.GetService(name)
-		if !ok {
-			return fmt.Errorf("service %q not found", name)
-		}
-		selected[name] = true
-		for _, dependency := range svc.Config.DependsOn {
-			if err := visit(dependency); err != nil {
-				return err
-			}
-		}
-		return nil
-	}
-	for _, name := range names {
-		if err := visit(name); err != nil {
-			return nil, err
-		}
-	}
-	return selected, nil
-}
-
-// Shutdown rejects new starts and stops every child process exactly once.
-func (m *Manager) Shutdown() error {
-	m.shuttingDown.Store(true)
-	m.stopListenerDiscovery()
-	m.stopAllStatusMonitors()
-	m.stopAllDetachedLogs()
-	m.actions.CancelActive()
-	names := make([]string, 0)
-	for _, svc := range m.Services() {
-		if svc.Config.StopOnExitEnabled() {
-			names = append(names, svc.Name)
-		}
-	}
-	err := m.ForceStopServices(names)
-	m.actions.Shutdown()
-	return err
-}
-
-// RestartService restarts a service and all transitive dependents.
-func (m *Manager) RestartService(name string) error {
-	// Hold reloadMu for the whole operation so a concurrent config reload
-	// cannot swap or remove services mid-restart out from under this plan.
-	m.reloadMu.Lock()
-	defer m.reloadMu.Unlock()
-	if _, ok := m.GetService(name); !ok {
-		return fmt.Errorf("service %q not found", name)
-	}
-
-	order, err := m.topologicalSort()
-	if err != nil {
-		return err
-	}
-	affectedSet := map[string]bool{name: true}
-	for _, dependent := range m.findDependents(name) {
-		if svc, ok := m.GetService(dependent); ok && svc.Status() != config.StatusStopped {
-			affectedSet[dependent] = true
-		}
-	}
-	var affected []string
-	for _, serviceName := range order {
-		if affectedSet[serviceName] {
-			affected = append(affected, serviceName)
-		}
-	}
-
-	// Stop dependents before their dependencies.
-	for i := len(affected) - 1; i >= 0; i-- {
-		if err := m.StopService(affected[i]); err != nil {
-			return fmt.Errorf("stop service %q: %w", affected[i], err)
-		}
-	}
-
-	// Start dependencies before their dependents.
-	for _, n := range affected {
-		if err := m.StartService(n); err != nil {
-			return fmt.Errorf("start service %q: %w", n, err)
-		}
-	}
-
-	return nil
-}
-
-// RestartAll restarts only services that were active when the operation began.
-func (m *Manager) RestartAll() error {
-	// See RestartService: block a concurrent reload for the whole operation.
-	m.reloadMu.Lock()
-	defer m.reloadMu.Unlock()
-	order, err := m.topologicalSort()
-	if err != nil {
-		return err
-	}
-	running := make(map[string]bool)
-	for _, name := range order {
-		if svc, ok := m.GetService(name); ok && svc.Status() != config.StatusStopped {
-			running[name] = true
-		}
-	}
-	for i := len(order) - 1; i >= 0; i-- {
-		if running[order[i]] {
-			if err := m.StopService(order[i]); err != nil {
-				return err
-			}
-		}
-	}
-	for _, name := range order {
-		if running[name] {
-			if err := m.StartService(name); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// GetAffectedServices returns the restart target followed by transitive dependents.
-func (m *Manager) GetAffectedServices(name string) []string {
-	order, err := m.topologicalSort()
-	if err != nil {
-		return []string{name}
-	}
-	set := map[string]bool{name: true}
-	for _, dependent := range m.findDependents(name) {
-		if svc, ok := m.GetService(dependent); ok && svc.Status() != config.StatusStopped {
-			set[dependent] = true
-		}
-	}
-	result := make([]string, 0, len(set))
-	for _, serviceName := range order {
-		if set[serviceName] {
-			result = append(result, serviceName)
-		}
-	}
-	return result
-}
-
-// topologicalSort orders services with Kahn's algorithm.
-func (m *Manager) topologicalSort() ([]string, error) {
-	cfg := m.configSnapshot()
-	graph := cfg.GetDependsOn()
-	names := cfg.ServiceNames()
-	inDegree := make(map[string]int, len(names))
-	dependents := make(map[string][]string, len(names))
-
-	for _, name := range names {
-		inDegree[name] = len(graph[name])
-		for _, dependency := range graph[name] {
-			dependents[dependency] = append(dependents[dependency], name)
-		}
-	}
-
-	var queue []string
-	for _, name := range names {
-		if inDegree[name] == 0 {
-			queue = append(queue, name)
-		}
-	}
-
-	var result []string
-	for len(queue) > 0 {
-		node := queue[0]
-		queue = queue[1:]
-		result = append(result, node)
-
-		for _, dependent := range dependents[node] {
-			inDegree[dependent]--
-			if inDegree[dependent] == 0 {
-				queue = append(queue, dependent)
-			}
-		}
-	}
-
-	if len(result) != len(names) {
-		return nil, errors.New("dependency cycle detected")
-	}
-
-	return result, nil
-}
-
-// findDependents returns every transitive dependent of a service.
-func (m *Manager) findDependents(name string) []string {
-	graph := m.configSnapshot().GetDependsOn()
-	var result []string
-	visited := make(map[string]bool)
-
-	var dfs func(current string)
-	dfs = func(current string) {
-		for svcName, deps := range graph {
-			for _, dep := range deps {
-				if dep == current && !visited[svcName] {
-					visited[svcName] = true
-					result = append(result, svcName)
-					dfs(svcName)
-				}
-			}
-		}
-	}
-
-	dfs(name)
-	return result
-}
-
-// monitorProcess drains output, observes completion, and applies recovery policy.
-func (m *Manager) monitorProcess(name string, svc *Service, pm *ProcessManager, cancelCh chan struct{}) {
-	// Drain process buffers frequently enough to keep the UI responsive.
-	ticker := time.NewTicker(200 * time.Millisecond)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-cancelCh:
-			m.drainProcessLogs(svc, pm)
-			return
-		case <-pm.Done():
-			m.drainProcessLogs(svc, pm)
-			waitErr := pm.Wait()
-			exitCode := pm.ExitCode()
-			m.mu.RLock()
-			hc := m.healthChecker
-			m.mu.RUnlock()
-			if hc != nil {
-				hc.StopMonitoring(name)
-			}
-			svc.lifecycleMu.Lock()
-			shouldEvaluate := false
-			if current, _ := svc.runtime(); current == pm && svc.Status() != config.StatusStopping {
-				svc.clearRuntime(pm)
-				svc.SetStatus(config.StatusStopped)
-				svc.SetPID(0)
-				svc.RecordExit(exitCode, waitErr)
-				shouldEvaluate = true
-				if m.successfulExit(svc.Config, exitCode) {
-					svc.AppendLog(fmt.Sprintf("[Kranz] Process exited · code %d", exitCode))
-				} else {
-					svc.AppendLog(fmt.Sprintf("[Kranz] Process failed · exit code %d", exitCode))
-				}
-			}
-			svc.lifecycleMu.Unlock()
-			if shouldEvaluate {
-				m.handleNaturalExit(name, svc, exitCode)
-			}
-			return
-		case <-ticker.C:
-			m.drainProcessLogs(svc, pm)
-		}
-	}
-}
-
-type listenerDiscoveryTarget struct {
-	service    *Service
-	leaderPID  int
-	generation uint64
-}
-
-func (m *Manager) ensureListenerDiscovery() {
-	m.discoveryMu.Lock()
-	defer m.discoveryMu.Unlock()
-	if m.listenerScanner == nil || m.discoveryCancel != nil || m.shuttingDown.Load() {
-		return
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	done := make(chan struct{})
-	m.discoveryCancel = cancel
-	m.discoveryDone = done
-	interval := m.listenerScanInterval
-	if interval <= 0 {
-		interval = 2 * time.Second
-	}
-	go func() {
-		defer close(done)
-		m.refreshDetectedPorts(ctx)
-		ticker := time.NewTicker(interval)
-		defer ticker.Stop()
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			case <-ticker.C:
-				m.refreshDetectedPorts(ctx)
-			}
-		}
-	}()
-}
-
-func (m *Manager) stopListenerDiscovery() {
-	m.discoveryMu.Lock()
-	cancel := m.discoveryCancel
-	done := m.discoveryDone
-	m.discoveryCancel = nil
-	m.discoveryDone = nil
-	m.discoveryMu.Unlock()
-	if cancel == nil {
-		return
-	}
-	cancel()
-	<-done
-}
-
-func (m *Manager) refreshDetectedPorts(ctx context.Context) {
-	m.discoveryMu.Lock()
-	scanner := m.listenerScanner
-	m.discoveryMu.Unlock()
-	if scanner == nil {
-		return
-	}
-
-	targets := make(map[string]listenerDiscoveryTarget)
-	for _, svc := range m.Services() {
-		if !svc.Config.PortDiscoveryEnabled() {
-			continue
-		}
-		leaderPID, generation, running := svc.discoveryTarget()
-		if running {
-			targets[svc.Name] = listenerDiscoveryTarget{
-				service: svc, leaderPID: leaderPID, generation: generation,
-			}
-		}
-	}
-	if len(targets) == 0 {
-		return
-	}
-
-	listeners, err := scanner.Snapshot(ctx)
-	if err != nil {
-		return
-	}
-	portsByService := make(map[string][]int, len(targets))
-	for _, listener := range listeners {
-		if !strings.EqualFold(listener.Protocol, "tcp") || listener.PID < 1 {
-			continue
-		}
-		for name, target := range targets {
-			if sameProcessGroup(target.leaderPID, listener.PID) {
-				portsByService[name] = append(portsByService[name], listener.Port)
-				break
-			}
-		}
-	}
-	for name, target := range targets {
-		target.service.updateDetectedPorts(target.generation, portsByService[name])
-	}
-}
-
-func (m *Manager) successfulExit(svc config.Service, exitCode int) bool {
-	if exitCode == 0 {
-		return true
-	}
-	for _, code := range svc.SuccessExitCodes {
-		if code == exitCode {
-			return true
-		}
-	}
-	return false
-}
-
-func (m *Manager) handleNaturalExit(name string, svc *Service, exitCode int) {
-	availability := svc.Config.Availability
-	success := m.successfulExit(svc.Config, exitCode)
-	restart := availability.Restart == "always" || (availability.Restart == "on_failure" && !success)
-	state := svc.GetState()
-	restartAllowed := availability.MaxRestarts == 0 || state.RestartCount < availability.MaxRestarts
-	if restart && svc.DesiredRunning() && !m.shuttingDown.Load() && restartAllowed {
-		attempt := svc.IncrementRestartCount()
-		backoff := availability.Backoff
-		if backoff <= 0 {
-			backoff = time.Second
-		}
-		svc.AppendLog(fmt.Sprintf("[Kranz] Restart scheduled · attempt %d · in %s", attempt, backoff))
-		go func() {
-			timer := time.NewTimer(backoff)
-			defer timer.Stop()
-			<-timer.C
-			if !svc.DesiredRunning() || m.shuttingDown.Load() {
-				return
-			}
-			if err := m.startService(context.Background(), name, true); err != nil {
-				svc.AppendLog("[Kranz] Automatic restart failed: " + err.Error())
-			}
-		}()
-		return
-	}
-	if restart && !restartAllowed {
-		svc.AppendLog(fmt.Sprintf("[Kranz] Restart limit reached (%d)", availability.MaxRestarts))
-	}
-	svc.SetDesiredRunning(false)
-	if availability.ExitOnEnd || (availability.Restart == "exit_on_failure" && !success) {
-		svc.AppendLog("[Kranz] Project stop requested by availability policy")
-		requestedCode := exitCode
-		if success {
-			requestedCode = 0
-		}
-		m.requestProjectExit(requestedCode)
-	}
-}
-
-func (m *Manager) requestProjectExit(code int) {
-	if !m.exitRequested.CompareAndSwap(false, true) {
-		return
-	}
-	m.exitCode.Store(int64(code))
-	go func() { _ = m.StopAll() }()
-}
-
-// ProjectExitRequested reports whether an availability policy requested the
-// whole Kranz session to terminate, together with its intended exit code.
-func (m *Manager) ProjectExitRequested() (bool, int) {
-	return m.exitRequested.Load(), int(m.exitCode.Load())
-}
-
-func (m *Manager) drainProcessLogs(svc *Service, pm *ProcessManager) {
-	for _, line := range pm.Stdout().Drain() {
-		svc.AppendLog(line)
-	}
-	for _, line := range pm.Stderr().Drain() {
-		svc.AppendLog(line)
-	}
-}
-
-// HasRunningServices reports whether any managed service is active or transitioning.
 func (m *Manager) HasRunningServices() bool {
 	m.mu.RLock()
 	defer m.mu.RUnlock()
@@ -1383,17 +286,7 @@ func (m *Manager) GetAllTags() []string {
 }
 
 // PortConflictError describes the verified owner of a required listening port.
-type PortConflictError struct {
-	Service      string
-	Port         int
-	PID          int
-	Process      string
-	Command      string
-	OwnerService string
-	External     bool
-}
 
-// Error returns a concise description of the conflicting listener.
 func (e *PortConflictError) Error() string {
 	return fmt.Sprintf("port %d is occupied by PID %d (%s)", e.Port, e.PID, e.Process)
 }
@@ -1401,59 +294,3 @@ func (e *PortConflictError) Error() string {
 // ManagedServiceForPID returns the Kranz service that owns pid. A service may
 // launch the actual listener as a child of its shell, so ownership is matched
 // by process group as well as by the recorded leader PID.
-func (m *Manager) ManagedServiceForPID(pid int) string {
-	if pid <= 0 {
-		return ""
-	}
-	for _, svc := range m.Services() {
-		leader := svc.PID()
-		if leader > 0 && sameProcessGroup(leader, pid) {
-			return svc.Name
-		}
-	}
-	return ""
-}
-
-// waitForReadiness blocks until readiness succeeds, times out, or is cancelled.
-func (m *Manager) waitForReadiness(ctx context.Context, name string, timeout time.Duration) bool {
-	svc, ok := m.GetService(name)
-	if !ok {
-		return false
-	}
-	if svc.Config.HealthCheck == nil || svc.Config.HealthCheck.Readiness == nil {
-		status := svc.Status()
-		return status == config.StatusRunning || status == config.StatusUnhealthy
-	}
-
-	m.mu.RLock()
-	hc := m.healthChecker
-	m.mu.RUnlock()
-
-	if hc == nil {
-		return true // Without a checker there is no readiness gate.
-	}
-
-	timer := time.NewTimer(timeout)
-	defer timer.Stop()
-	ticker := time.NewTicker(100 * time.Millisecond)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-ctx.Done():
-			return false
-		case <-timer.C:
-			return false
-		case <-ticker.C:
-		}
-		if m.shuttingDown.Load() {
-			return false
-		}
-		if svc, ok := m.GetService(name); !ok || svc.Status() == config.StatusStopped {
-			return false
-		}
-		health := hc.GetHealth(name)
-		if health != nil && health.IsReady() {
-			return true
-		}
-	}
-}
